requirements.txt
Django==2.1.4
requests==2.13.0
django-cors-headers==1.2.0
django-debug-toolbar==1.8
graphene-django==2.2.0
django-filter==2.0.0
django-rest-auth==0.9.3
boto3==1.9.104
django-environ==0.4.5